Add support for Solaris
This required a serious restructuring because it turns out that the Go
standard library does not actually implement the net.Interface stuff on
Solaris. So not only do we have to get device statistics through kstats
instead of /proc/net/dev (which I saw coming from the start), but we
have to inventory network devices themselves in a system dependent way
(on Solaris we use cgo to call getifaddrs()).