Add -a to adapt bandwidth units to current volume
The new -a switch makes netvolmon switch between KB/s, MB/s, and even
GB/s depending on how much traffic volume you have in the current
reporting period. This is kind of a peculiar feature and I'm not sure
it's a good idea, but I'm experimenting here.

If you're monitoring multiple interfaces, the current implementation
picks units separately for each of them. This may be considered a
feature or a bug depending on your perspective.